# Backend/swagger.yaml
openapi: 3.0.3
info:
  title: Blood4U API
  description: |
    Production-ready REST API for the Blood4U Blood Donation Management System.
    Provides endpoints for authentication, donor management, blood requests, blood banks,
    community events, and real-time chat.
  version: 2.0.0
  contact:
    name: Blood4U Support
servers:
  - url: http://localhost:3000/api
    description: Local Development
  - url: https://your-production-domain.com/api
    description: Production
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    Donor:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        fullName:
          type: string
        bloodGroup:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        city:
          type: string
        available:
          type: boolean
        donationCount:
          type: integer
        lastDonationDate:
          type: string
          format: date-time
        badge:
          type: string
          enum: [Bronze, Silver, Gold]
        isResting:
          type: boolean
    BloodRequest:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        bloodGroup:
          type: string
        city:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, fulfilled]
        urgent:
          type: boolean
        matchedDonors:
          type: array
          items:
            type: string
    BloodBank:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        city:
          type: string
        contact:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        location:
          type: object
          properties:
            type:
              type: string
            coordinates:
              type: array
              items:
                type: number
security:
  - cookieAuth: []
paths:
...
  /donors:
    get:
      summary: Get all donors (filterable)
      tags: [Donors]
      parameters:
        - in: query
          name: bloodGroup
          schema:
            type: string
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: available
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: List of donors

    post:
      summary: Register as a blood donor
      tags: [Donors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Donor'
      responses:
        '201':
          description: Donor registered
        '409':
          description: Already a donor
...
  /requests:
    get:
      summary: Get blood requests (role-based visibility)
      tags: [Requests]
      responses:
        '200':
          description: List of requests

    post:
      summary: Create a new blood request
      tags: [Requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BloodRequest'
      responses:
        '201':
          description: Request created and donors matched
...
  /banks:
    get:
      summary: Get all blood banks
      tags: [Blood Banks]
      responses:
        '200':
          description: List of blood banks

    post:
      summary: Register a blood bank
      tags: [Blood Banks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BloodBank'
      responses:
        '201':
          description: Blood bank submitted for approval
...